Move cec.js file to the cec directory and implement logging by using streams
